travis: Use rustup to install alternate targets.
This seems likely to be more reliable than the external script.
Whether it's easier to read or not depends on how much you
like sh crammed into travis.yml...

The rustup tool treats trying to install the host triplet
as a new target as an error, so we have to run the installer
only if the architecture doesn't match the host.

Also removed the reduntant echo of the matrix environment
variables. They are already printed in the export section
of the log near the start. There's no need to write them
out again when invoking cargo.